Cambios nuevos controladores y DNI

+++ TalentoInterno.CORE/Core/DTOs/ColaboradorSkillDto.cs
@@ -5,7 +5,7 @@ public class ColaboradorSkillDto
     public int ColaboradorId { get; set; }
     public int SkillId { get; set; }
     public string? SkillNombre { get; set; }
-    public byte NivelId { get; set; }
+    public int NivelId { get; set; }
     public string? NivelNombre { get; set; }
     public decimal? AniosExp { get; set; }
 }
