+++ TalentoInterno.CORE/Core/DTOs/UrgenciaDto.cs
@@ -2,6 +2,6 @@ namespace TalentoInterno.CORE.Core.DTOs;
 
 public class UrgenciaDto
 {
-    public byte UrgenciaId { get; set; }
+    public int UrgenciaId { get; set; }
     public string Nombre { get; set; } = null!;
 }
