+++ TalentoInterno.CORE/Core/Services/ColaboradorSkillService.cs
@@ -149,8 +149,8 @@ public class ColaboradorSkillService : IColaboradorSkillService
             {
                 var skillColab = colab.ColaboradorSkill.FirstOrDefault(s => s.SkillId == req.SkillId);
 
-                byte nivelReq = req.NivelDeseado;
-                byte nivelAct = skillColab?.NivelId ?? 0;
+                int nivelReq = req.NivelDeseado;
+                int nivelAct = skillColab?.NivelId ?? 0;
 
                 double brecha = 0;
                 if (nivelReq > nivelAct)
